refactor: move KotlinCleanup to common module
