+++ AnkiDroid/src/main/java/com/ichi2/anki/AnkiDroidApp.kt
@@ -38,6 +38,7 @@ import anki.collection.OpChanges
 import com.ichi2.anki.CrashReportService.sendExceptionReport
 import com.ichi2.anki.analytics.UsageAnalytics
 import com.ichi2.anki.browser.SharedPreferencesLastDeckIdRepository
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.common.utils.isRunningAsUnitTest
 import com.ichi2.anki.contextmenu.AnkiCardContextMenu
 import com.ichi2.anki.contextmenu.CardBrowserContextMenu
@@ -58,7 +59,6 @@ import com.ichi2.compat.CompatHelper
 import com.ichi2.libanki.ChangeManager
 import com.ichi2.utils.AdaptionUtil
 import com.ichi2.utils.ExceptionUtil
-import com.ichi2.utils.KotlinCleanup
 import com.ichi2.utils.LanguageUtil
 import com.ichi2.utils.Permissions
 import com.ichi2.widget.cardanalysis.CardAnalysisWidget
