+++ AnkiDroid/src/main/java/com/ichi2/anki/MetaDB.kt
@@ -7,11 +7,11 @@ import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteException
 import androidx.annotation.WorkerThread
 import androidx.core.database.sqlite.transaction
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.model.WhiteboardPenColor
 import com.ichi2.anki.model.WhiteboardPenColor.Companion.default
 import com.ichi2.anki.reviewer.CardSide
 import com.ichi2.libanki.DeckId
-import com.ichi2.utils.KotlinCleanup
 import com.ichi2.widget.SmallWidgetStatus
 import timber.log.Timber
