+++ AnkiDroid/src/main/java/com/ichi2/anki/multimediacard/fields/ImageField.kt
@@ -22,9 +22,9 @@ package com.ichi2.anki.multimediacard.fields
 import android.net.Uri
 import androidx.annotation.CheckResult
 import androidx.annotation.VisibleForTesting
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.annotations.NeedsTest
 import com.ichi2.libanki.Collection
-import com.ichi2.utils.KotlinCleanup
 import org.jsoup.Jsoup
 import timber.log.Timber
 import java.io.File
