+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionOperations.kt
@@ -17,9 +17,9 @@
 package com.ichi2.async
 
 import com.ichi2.anki.CardTemplateNotetype
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.NotetypeJson
-import com.ichi2.utils.KotlinCleanup
 import timber.log.Timber
 
 /**
