+++ AnkiDroid/src/main/java/com/ichi2/utils/HashUtil.kt
@@ -15,6 +15,7 @@
  ****************************************************************************************/
 package com.ichi2.utils
 
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import java.util.HashMap
 import java.util.HashSet
