+++ AnkiDroid/src/main/java/com/ichi2/utils/JSONArray.kt
@@ -41,6 +41,7 @@
  */
 package com.ichi2.utils
 
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import org.json.JSONArray
 import org.json.JSONObject
