+++ AnkiDroid/src/main/java/com/ichi2/utils/UniqueArrayList.kt
@@ -18,6 +18,7 @@
 
 package com.ichi2.utils
 
+import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import org.apache.commons.collections4.list.SetUniqueList
 import java.util.Arrays
 import java.util.Spliterator
